refactor: swap index & marketplace
