--- pages/marketplace.js
@@ -1,13 +1,0 @@
-import VehicleFeed from '@components/VehicleFeed';
-import React from 'react';
-
-const Marketplace = () => {
-    return (
-        <div>
-            <div className='background'/>
-            <VehicleFeed/>
-        </div>
-    )
-};
-
-export default Marketplace;
